Splitting the Concat builder.
Code review commit taking into account Jakub's remarks.

The concat builder was used for two different purposes:
 - concatenating text files together
 - copying binary files
Additionally, it was able to handle another case, which
will probably not be very useful:
 - concatenating several binary files together
Because this third case is not very common, and makes the code
of Concat a bit more complex, this commit is splitting the two
main purposes of the Concat builder into 2 different builders
(Concat for text files and Copy for binary files), and removes
support for the third edge case.