Adding parameter to exclude packages files excluded from global sourceFiles
This commit adds an extra configuration parameter for atpackager: `onlySourceFiles`. If set to true, it will prevent packages definitions from adding files that are globally excluded by the `sourceFiles` option.
The default value for `onlySourceFiles` is `false` for backward compatibility.

The reason for the change is a performance gain in the construction of the packages: looking for files in a list that has already been built when initializing atpackager according to the `sourceFiles` option is much faster than rebuilding the list of available files.

Close #15.