Updating the test to check whether a template is already compiled.
This commit follows the following change in Aria Templates:
https://github.com/ariatemplates/ariatemplates/commit/b42d5caee913265da424406ec620865ebe6020e8

The detection code is copied from GeneralTplPreprocessor.

It is not possible to directly call GeneralTplPreprocessor (instead of
copying it) because the allDeps parameter passed to the
template generator is not adapted to precompilation in that case
(it leads to missing dependencies when executing the template).
Of course, we could change the code in Aria Templates, but it would
then require extra conditions to keep atpackager compatible with all
Aria Templates versions (as it is currently).

Close #21